+++ server/lib/parties/party-service.test.ts
@@ -8,8 +8,7 @@ import {
 import { NotificationType } from '../../../common/notifications'
 import { asMockedFunction } from '../../../common/testing/mocks'
 import { makeSbUserId, SbUser } from '../../../common/users/sb-user'
-import { findChannelsByName, FullChannelInfo } from '../chat/chat-models'
-import { toBasicChannelInfo } from '../chat/chat-service'
+import { findChannelsByName, FullChannelInfo, toBasicChannelInfo } from '../chat/chat-models'
 import { GameplayActivityRegistry } from '../games/gameplay-activity-registry'
 import { MatchmakingServiceError } from '../matchmaking/matchmaking-service-error'
 import NotificationService from '../notifications/notification-service'
